MCP/source-120125/vertex-memory-bank-mcp-main/requirements.txt
# Core MCP and Memory Bank dependencies
mcp[cli]>=1.0.0
google-cloud-aiplatform>=1.118.0
vertexai>=1.0.0
google-genai>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Additional dependencies for Gemini integration and ADK
google-genai>=0.1.0
google-adk>=1.15.1

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0